Move thread search into the quick palette
