+++ apps/app/src/components/commands/CommandPalette.tsx
@@ -7,6 +7,7 @@ import {
   useState,
 } from "react";
 import type { KeyboardEvent as ReactKeyboardEvent } from "react";
+import { useNavigate } from "react-router-dom";
 import { Dialog, DialogContent, DialogTitle } from "@bb/shared-ui/dialog";
 import { Icon } from "@bb/shared-ui/icon";
 import { cn } from "@bb/shared-ui/lib/utils";
@@ -34,8 +35,13 @@ import {
 import { buildPluginPaletteActions } from "@/lib/command-palette/palette-plugin-actions";
 import { getPluginSlotSnapshot } from "@/lib/plugin-slots";
 import { getActiveThreadPanelOpener } from "@/components/plugin/plugin-thread-panel-navigation";
+import { getThreadRoutePath } from "@/lib/route-paths";
+import {
+  ThreadPaletteResults,
+  type ThreadPaletteNavigationItem,
+} from "./ThreadPaletteResults";
 
-const PALETTE_PLACEHOLDER = "Search commands";
+type PaletteMode = "commands" | "threads";
 
 export interface CommandPaletteProps {
   /** The surface's thread and project, handed to plugin rows. */
@@ -48,6 +54,7 @@ export interface CommandPaletteProps {
  * Mounted once by `AppLayout` and opened by `palette.open`.
  */
 export function CommandPalette({ threadId, projectId }: CommandPaletteProps) {
+  const navigate = useNavigate();
   const runner = useAppCommandRunner();
   const shortcuts = useAppCommandShortcuts(PALETTE_COMMAND_IDS);
   const listId = useId();
@@ -56,6 +63,9 @@ export function CommandPalette({ threadId, projectId }: CommandPaletteProps) {
   const [open, setOpen] = useState(false);
   const [query, setQuery] = useState("");
   const [actions, setActions] = useState<readonly PaletteAction[]>([]);
+  const [threadItems, setThreadItems] = useState<
+    readonly ThreadPaletteNavigationItem[]
+  >([]);
   const [highlightedIndex, setHighlightedIndex] = useState(0);
   const [recents, setRecents] = useState<readonly string[]>(() =>
     readPaletteRecents(),
@@ -63,14 +73,10 @@ export function CommandPalette({ threadId, projectId }: CommandPaletteProps) {
   // Where availability, dispatch, and focus-on-close all point.
   const openTargetRef = useRef<EventTarget | null>(null);
   // Set when a row is chosen, read once focus has been restored.
-  const pendingActionRef = useRef<PaletteAction | null>(null);
+  const pendingRunRef = useRef<(() => void) | null>(null);
 
-  useAppCommandHandler("palette.open", (invocation) => {
-    const target =
-      invocation.target ??
-      (typeof document === "undefined" ? null : document.activeElement);
-    openTargetRef.current = target;
-    setActions([
+  const buildActions = useCallback(
+    (target: EventTarget | null) => [
       ...buildAppCommandActions({
         target,
         isCommandAvailable: runner.isCommandAvailable,
@@ -83,20 +89,55 @@ export function CommandPalette({ threadId, projectId }: CommandPaletteProps) {
         projectId,
         openThreadPanel: getActiveThreadPanelOpener(),
       }),
-    ]);
-    setQuery("");
-    setHighlightedIndex(0);
-    setOpen(true);
+    ],
+    [
+      projectId,
+      runner.dispatch,
+      runner.isCommandAvailable,
+      shortcuts,
+      threadId,
+    ],
+  );
+
+  const openPalette = useCallback(
+    (mode: PaletteMode, target: EventTarget | null) => {
+      openTargetRef.current = target;
+      setActions(buildActions(target));
+      setThreadItems([]);
+      setQuery(mode === "commands" ? ">" : "");
+      setHighlightedIndex(0);
+      setOpen(true);
+    },
+    [buildActions],
+  );
+
+  useAppCommandHandler("palette.open", (invocation) => {
+    const target =
+      invocation.target ??
+      (typeof document === "undefined" ? null : document.activeElement);
+    openPalette("commands", target);
     return true;
   });
 
-  const ranked = useMemo(
-    () => rankPaletteActions({ actions, query, recentIds: recents }),
-    [actions, query, recents],
+  useAppCommandHandler("thread.search", (invocation) => {
+    const target =
+      invocation.target ??
+      (typeof document === "undefined" ? null : document.activeElement);
+    openPalette("threads", target);
+    return true;
+  });
+
+  const mode: PaletteMode = query.startsWith(">") ? "commands" : "threads";
+  const modeQuery = mode === "commands" ? query.slice(1) : query;
+  const rankedCommands = useMemo(
+    () => rankPaletteActions({ actions, query: modeQuery, recentIds: recents }),
+    [actions, modeQuery, recents],
   );
+  const resultCount =
+    mode === "commands" ? rankedCommands.length : threadItems.length;
   // Typing can shrink the list under the selection.
   const activeIndex =
-    ranked.length === 0 ? -1 : Math.min(highlightedIndex, ranked.length - 1);
+    resultCount === 0 ? -1 : Math.min(highlightedIndex, resultCount - 1);
 
   /**
    * Focus stays in the search field, so nothing scrolls the highlighted row
@@ -114,34 +155,63 @@ export function CommandPalette({ threadId, projectId }: CommandPaletteProps) {
   }, [activeIndex]);
 
   const chooseAction = useCallback((action: PaletteAction) => {
-    pendingActionRef.current = action;
+    if (action.id === "app:thread.search") {
+      setQuery("");
+      setHighlightedIndex(0);
+      if (listRef.current !== null) listRef.current.scrollTop = 0;
+      return;
+    }
+    pendingRunRef.current = action.run;
     setRecents((current) => recordPaletteRecent(current, action.id));
     setOpen(false);
   }, []);
+
+  const chooseThread = useCallback(
+    (item: ThreadPaletteNavigationItem) => {
+      pendingRunRef.current = () => {
+        void navigate(
+          getThreadRoutePath({
+            projectId: item.projectId,
+            threadId: item.threadId,
+          }),
+          item.messageSeq === null
+            ? undefined
+            : {
+                state: {
+                  searchMessageSeq: item.messageSeq,
+                  searchThreadId: item.threadId,
+                },
+              },
+        );
+      };
+      setOpen(false);
+    },
+    [navigate],
+  );
 
   /**
    * Restore focus before running, so a command that focuses something does not
    * have it taken back by the dialog's own restoration a tick later.
    */
   const handleCloseAutoFocus = useCallback((event: Event) => {
-    const pending = pendingActionRef.current;
-    pendingActionRef.current = null;
+    const pending = pendingRunRef.current;
+    pendingRunRef.current = null;
     const target = openTargetRef.current;
     if (target instanceof HTMLElement && target.isConnected) {
       event.preventDefault();
       target.focus({ preventScroll: true });
     }
-    pending?.run();
+    pending?.();
   }, []);
 
   const handleKeyDown = useCallback(
     (event: ReactKeyboardEvent<HTMLInputElement>) => {
-      if (ranked.length === 0) return;
+      if (resultCount === 0) return;
       if (event.key === "ArrowDown") {
         event.preventDefault();
         scrollOnNextHighlightRef.current = true;
         setHighlightedIndex((current) =>
-          current + 1 >= ranked.length ? 0 : current + 1,
+          current + 1 >= resultCount ? 0 : current + 1,
         );
         return;
       }
@@ -149,7 +219,7 @@ export function CommandPalette({ threadId, projectId }: CommandPaletteProps) {
         event.preventDefault();
         scrollOnNextHighlightRef.current = true;
         setHighlightedIndex((current) =>
-          current <= 0 ? ranked.length - 1 : current - 1,
+          current <= 0 ? resultCount - 1 : current - 1,
         );
         return;
       }
@@ -162,18 +232,38 @@ export function CommandPalette({ threadId, projectId }: CommandPaletteProps) {
       if (event.key === "End") {
         event.preventDefault();
         scrollOnNextHighlightRef.current = true;
-        setHighlightedIndex(ranked.length - 1);
+        setHighlightedIndex(resultCount - 1);
         return;
       }
       if (event.key === "Enter") {
-        const choice = ranked[activeIndex];
-        if (choice === undefined) return;
         event.preventDefault();
-        chooseAction(choice.action);
+        if (mode === "commands") {
+          const choice = rankedCommands[activeIndex];
+          if (choice !== undefined) chooseAction(choice.action);
+          return;
+        }
+        const choice = threadItems[activeIndex];
+        if (choice !== undefined) chooseThread(choice);
       }
     },
-    [activeIndex, chooseAction, ranked],
+    [
+      activeIndex,
+      chooseAction,
+      chooseThread,
+      mode,
+      rankedCommands,
+      resultCount,
+      threadItems,
+    ],
   );
+
+  const activeDescendant =
+    activeIndex === -1
+      ? undefined
+      : mode === "commands"
+        ? `${optionIdPrefix}-${activeIndex}`
+        : threadItems[activeIndex]?.optionId;
+  const inputLabel = mode === "commands" ? "Search commands" : "Search threads";
 
   return (
     <Dialog open={open} onOpenChange={setOpen}>
@@ -184,7 +274,9 @@ export function CommandPalette({ threadId, projectId }: CommandPaletteProps) {
         onCloseAutoFocus={handleCloseAutoFocus}
         data-testid="command-palette"
       >
-        <DialogTitle className="sr-only">Quick palette</DialogTitle>
+        <DialogTitle className="sr-only">
+          {mode === "commands" ? "Quick palette" : "Search threads"}
+        </DialogTitle>
         <div className="flex items-center gap-2 border-b px-3">
           <Icon
             name="Search"
@@ -196,19 +288,17 @@ export function CommandPalette({ threadId, projectId }: CommandPaletteProps) {
             role="combobox"
             aria-expanded
             aria-controls={listId}
-            aria-activedescendant={
-              activeIndex === -1
-                ? undefined
-                : `${optionIdPrefix}-${activeIndex}`
-            }
-            aria-label={PALETTE_PLACEHOLDER}
+            aria-activedescendant={activeDescendant}
+            aria-label={inputLabel}
             autoComplete="off"
             spellCheck={false}
             className="h-11 w-full bg-transparent text-sm outline-none placeholder:text-muted-foreground"
-            placeholder={PALETTE_PLACEHOLDER}
+            placeholder={inputLabel}
             value={query}
             onChange={(event) => {
-              setQuery(event.target.value);
+              const nextQuery = event.target.value;
+              setQuery(nextQuery);
+              if (!nextQuery.startsWith(">")) setThreadItems([]);
               setHighlightedIndex(0);
               // `activeIndex` may not change, so the effect above cannot do
               // this: send the scrolled container back to the first row.
@@ -221,15 +311,17 @@ export function CommandPalette({ threadId, projectId }: CommandPaletteProps) {
           ref={listRef}
           id={listId}
           role="listbox"
-          aria-label="Commands"
+          aria-label={
+            mode === "commands" ? "Commands" : "Thread search results"
+          }
           className="max-h-[min(24rem,50dvh)] overflow-y-auto p-1"
         >
-          {ranked.length === 0 ? (
+          {mode === "commands" && rankedCommands.length === 0 ? (
             <p className="px-2 py-6 text-center text-sm text-muted-foreground">
               No matching commands
             </p>
-          ) : (
-            ranked.map((entry, index) => (
+          ) : mode === "commands" ? (
+            rankedCommands.map((entry, index) => (
               <PaletteRow
                 key={entry.action.id}
                 entry={entry}
@@ -239,6 +331,15 @@ export function CommandPalette({ threadId, projectId }: CommandPaletteProps) {
                 onSelect={() => chooseAction(entry.action)}
               />
             ))
+          ) : (
+            <ThreadPaletteResults
+              activeIndex={activeIndex}
+              onActiveIndexChange={setHighlightedIndex}
+              onNavigationItemsChange={setThreadItems}
+              onSelect={chooseThread}
+              optionIdPrefix={optionIdPrefix}
+              query={modeQuery}
+            />
           )}
         </div>
       </DialogContent>
