--- apps/app/src/components/sidebar/sidebarThreadSearch.ts
@@ -1,79 +1,0 @@
-import type { RefObject } from "react";
-import type { ThreadSearchMatch } from "@bb/server-contract";
-
-export const SIDEBAR_THREAD_SEARCH_LISTBOX_ID =
-  "bb-sidebar-thread-search-results";
-
-export interface SidebarThreadSearchNavigationItem {
-  id: string;
-  optionId: string;
-  projectId: string;
-  threadId: string;
-  /**
-   * Event sequence of the matched message, so selecting the result can scroll
-   * to that message in the thread. Null when the match is a title or the row is
-   * a recent (no-query) entry with no message match.
-   */
-  messageSeq: number | null;
-}
-
-export interface SidebarThreadSearchInputController {
-  activeDescendantId: string | undefined;
-  inputRef: RefObject<HTMLInputElement | null>;
-  isActive: boolean;
-  onActivate: () => void;
-  onClose: () => void;
-  onQueryChange: (query: string) => void;
-  query: string;
-}
-
-export interface SidebarThreadSearchPanelController {
-  activeIndex: number;
-  isActive: boolean;
-  onActiveIndexChange: (index: number) => void;
-  onNavigationItemsChange: (
-    items: readonly SidebarThreadSearchNavigationItem[],
-  ) => void;
-  onSelectItem: (item: SidebarThreadSearchNavigationItem) => void;
-  query: string;
-}
-
-/**
- * The sidebar-wide key handler only owns keys typed in the search field or on a
- * result row. Every other sidebar control keeps its own key behavior.
- */
-export function isThreadSearchKeyboardEventTarget(
-  target: EventTarget | null,
-  input: HTMLInputElement | null,
-): boolean {
-  if (!(target instanceof HTMLElement)) {
-    return false;
-  }
-  if (target === input) {
-    return true;
-  }
-  return target.closest('[role="option"]') !== null;
-}
-
-export function getSidebarThreadSearchOptionId(rowId: string): string {
-  return `${SIDEBAR_THREAD_SEARCH_LISTBOX_ID}-option-${rowId}`;
-}
-
-export function isSidebarThreadTitleMatch(match: ThreadSearchMatch): boolean {
-  return match.sourceKind === "title" || match.sourceKind === "title_fallback";
-}
-
-export function haveSameSidebarThreadSearchNavigationItems(
-  left: readonly SidebarThreadSearchNavigationItem[],
-  right: readonly SidebarThreadSearchNavigationItem[],
-): boolean {
-  if (left.length !== right.length) {
-    return false;
-  }
-  return left.every(
-    (item, index) =>
-      item.id === right[index]?.id &&
-      item.optionId === right[index]?.optionId &&
-      item.messageSeq === right[index]?.messageSeq,
-  );
-}
